feat: Add delivery type filter to customize report
This commit adds a new filter option for delivery type in the customize
report interface. It includes a dropdown menu allowing users to select
either "Pickup" or "Delivery" as the delivery type. The corresponding
changes are made in the components `CustomizedReportsTable.vue` and
`ProductCategoryReport.vue`, as well as the pages
`customize-report/index.vue` and `product-category-report/index.vue`.
This feature provides more flexibility in generating customized reports
based on the delivery type.